backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["benchprincess <realjiwon98@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
djangorestframework = "^3.15.1"
mysqlclient = "^2.2.4"
djangorestframework-simplejwt = "^5.3.1"
django-allauth = "^0.61.1"
faker = "^24.4.0"
django-extensions = "^3.2.3"
ipython = "^8.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
